refactor: 모드 분기 훅 분리 정리
