+++ src/layout/RootLayout.tsx
@@ -1,20 +1,7 @@
-import { useEffect } from 'react';
-
-import { Outlet, useLocation } from 'react-router-dom';
-
-import { ROUTES } from '@/routes/paths';
+import { Outlet } from 'react-router-dom';
 import { useScrollToTop } from '@/shared/hooks/useScrollToTop';
 
-import { IS_CLIENT_DETECTION_MODE } from '@pages/generate/constants/curationDetectionMode';
-import { OBJ365_MODEL_PATH } from '@pages/generate/constants/detection';
-import { preloadONNXModel } from '@pages/generate/hooks/useOnnxModel';
-
-const GENERATE_WARMUP_PATHS = [
-  ROUTES.GENERATE,
-  ROUTES.GENERATE_RESULT,
-  ROUTES.GENERATE_START,
-  ROUTES.IMAGE_SETUP,
-];
+import { useGenerateWarmup } from '@pages/generate/hooks/useGenerateWarmup';
 
 function RootLayout() {
   // 라우트/쿼리/해시/키 변화와 초기 마운트 시 스크롤 최상단으로 이동
@@ -25,23 +12,6 @@ function RootLayout() {
       <Outlet />
     </div>
   );
-}
-
-function useGenerateWarmup() {
-  const location = useLocation();
-
-  useEffect(() => {
-    if (!IS_CLIENT_DETECTION_MODE) return;
-
-    const pathname = location.pathname;
-    const shouldWarmup = GENERATE_WARMUP_PATHS.some(
-      (prefix) => pathname === prefix || pathname.startsWith(`${prefix}/`)
-    );
-
-    if (!shouldWarmup) return;
-
-    preloadONNXModel(OBJ365_MODEL_PATH).catch(() => undefined);
-  }, [location.pathname]);
 }
 
 export default RootLayout;
