+++ src/pages/generate/hooks/useStartPageModelPreload.ts
@@ -1,0 +1,18 @@
+import { useEffect } from 'react';
+
+import { IS_CLIENT_DETECTION_ENABLED } from '@pages/generate/constants/curationDetectionMode';
+import { OBJ365_MODEL_PATH } from '@pages/generate/constants/detection';
+
+import { preloadONNXModel } from './useOnnxModel';
+
+const useStartPageModelPreloadClient = () => {
+  useEffect(() => {
+    preloadONNXModel(OBJ365_MODEL_PATH).catch(() => undefined);
+  }, []);
+};
+
+const useStartPageModelPreloadServer = () => {};
+
+export const useStartPageModelPreload = IS_CLIENT_DETECTION_ENABLED
+  ? useStartPageModelPreloadClient
+  : useStartPageModelPreloadServer;
