+++ src/pages/generate/pages/result/ResultPage.tsx
@@ -11,7 +11,7 @@ import type {
 import { createImageDetailPlaceholder } from '@/pages/mypage/utils/resultNavigation';
 
 import Loading from '@components/loading/Loading';
-import { IS_CLIENT_DETECTION_MODE } from '@pages/generate/constants/curationDetectionMode';
+import { IS_CLIENT_DETECTION_ENABLED } from '@pages/generate/constants/curationDetectionMode';
 import { useABTest } from '@pages/generate/hooks/useABTest';
 import { useGetResultDataQuery } from '@pages/generate/hooks/useGenerate';
 import { useCurationStore } from '@pages/generate/stores/useCurationStore';
@@ -216,7 +216,7 @@ const ResultPage = () => {
           <GeneratedImgA
             result={result}
             onCurrentImgIdChange={setCurrentImgId}
-            shouldInferHotspots={IS_CLIENT_DETECTION_MODE}
+            shouldInferHotspots={IS_CLIENT_DETECTION_ENABLED}
             userProfile={forwardedUserProfile}
             detectionCache={forwardedDetectionMap ?? undefined}
             isSlideCountLoading={isSlideCountLoading}
@@ -225,7 +225,7 @@ const ResultPage = () => {
           <GeneratedImgB
             result={result}
             onCurrentImgIdChange={setCurrentImgId}
-            shouldInferHotspots={IS_CLIENT_DETECTION_MODE}
+            shouldInferHotspots={IS_CLIENT_DETECTION_ENABLED}
             detectionCache={forwardedDetectionMap ?? undefined}
           />
         )}
